// src/shared/types/chapters.enum.ts
  export enum ChaptersEnum {
  CHESS_GAME_1 = 'CHESS_GAME_1',       // 1chessGame.widget.tsx
  QUEST_2 = 'QUEST_2',                // 2quest.widget.tsx
  PUZZLE_3 = 'PUZZLE_3',              // 3puzzle.widget.tsx
  WAY_4 = 'WAY_4',                    // 4way.widget.tsx
  MAGIC_5 = 'MAGIC_5',                // 5magic.widget.tsx
  GARDEN_CHANCE_6 = 'GARDEN_CHANCE_6', // 6gardenChance.widget.tsx
  WEDDING_PROPOSITION_7 = 'WEDDING_PROPOSITION_7' // 7weddingProposition.widget.tsx
}